requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
discord==2.3.2
discord.py==2.3.2
ffmpeg==1.4
frozenlist==1.4.1
gTTS==2.5.1
idna==3.7
multidict==6.0.5
pillow==10.3.0
pycparser==2.22
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.90
requests==2.31.0
six==1.16.0
urllib3==2.2.1
yarl==1.9.4
